Dev mpi 2

+++ windse/ParameterManager.py
@@ -36,20 +36,26 @@ if main_file != "sphinx-build":
 
 ### THis is a special class that allows prints to go to file and terminal
 class Logger(object):
-    def __init__(self,filename,std):
+    def __init__(self, filename, std, rank):
         self.__dict__ = std.__dict__.copy() 
         self.terminal = std
-        self.log = open(filename, "a")
-        self.log.seek(0)
-        self.log.truncate()
+        self.rank = rank
+        if self.rank == 0:
+            self.log = open(filename, "a")
+            self.log.seek(0)
+            self.log.truncate()
 
     def write(self, message):
         self.terminal.write(message)
-        self.log.write(message)  
+
+        if self.rank == 0:
+            self.log.write(message)  
 
     def flush(self):
         self.terminal.flush()
-        self.log.flush()
+
+        if self.rank == 0:
+            self.log.flush()
         pass   
 
     def isatty(self):
@@ -67,6 +73,11 @@ class Parameters(dict):
         self.windse_path = os.path.dirname(os.path.realpath(__file__))
         self.defaults = yaml.load(open(self.windse_path+"/default_parameters.yaml"),Loader=yaml.SafeLoader)
         self.update(self.defaults)
+
+        # Create an MPI communicator and initialize rank and num_procs 
+        self.comm = MPI.comm_world
+        self.rank = self.comm.Get_rank()
+        self.num_procs = self.comm.Get_size()
 
     def TerminalUpdate(self,dic,keys,value):
         if len(keys) > 1:
@@ -178,14 +189,36 @@ class Parameters(dict):
         self.folder = self.output_folder+self.name+"/"
         self["general"]["folder"] = self.folder
 
-        ### Make sure folder exists ###
-        if not os.path.exists(self.folder): os.makedirs(self.folder)
-        if not os.path.exists(self.folder+"input_files/"): os.makedirs(self.folder+"input_files/")
+        # Create all needed directories ahead of time
+        if self.rank == 0:
+            # Try to create the parent folder
+            os.makedirs(self.folder, exist_ok=True)
+
+            # Try to create all sub folders within this parent
+            subfolder_list = ['data',
+                              'functions',
+                              'input_files',
+                              'mesh',
+                              'plots',
+                              'timeSeries',
+                              'profiling']
+
+            for sub in subfolder_list:
+                os.makedirs('%s/%s' % (self.folder, sub), exist_ok=True)
+
+        # Wait until rank 0 has created the directory structure
+        self.comm.barrier()
+
+        # ### Make sure folder exists ###
+        # comm = MPI.comm_world
+        # rank = comm.Get_rank()
+        # if not os.path.exists(self.folder) and rank == 0: os.makedirs(self.folder)
+        # if not os.path.exists(self.folder+"input_files/") and rank == 0: os.makedirs(self.folder+"input_files/")
         
         ### Setup the logger ###
         self.log = self.folder+"log.txt"
-        sys.stdout = Logger(self.log,sys.stdout)
-        sys.stderr = Logger(self.log,sys.stderr)
+        sys.stdout = Logger(self.log, sys.stdout, self.rank)
+        sys.stderr = Logger(self.log, sys.stderr, self.rank)
 
         ### Copy params file to output folder ###
         if isinstance(loc,str):
@@ -250,7 +283,7 @@ class Parameters(dict):
 
         if file is None:
             ### Make sure the folder exists
-            if not os.path.exists(self.folder+subfolder): os.makedirs(self.folder+subfolder)
+            if not os.path.exists(self.folder+subfolder) and self.rank == 0: os.makedirs(self.folder+subfolder)
 
             if filetype == "pvd":
                 file_string = self.folder+subfolder+filename+".pvd"
@@ -286,8 +319,7 @@ class Parameters(dict):
 
         """
         ### Check Processor ###
-        rank = 0
-        if rank == 0:
+        if self.rank == 0:
             ### Check if tab length has been overridden
             if tab is None:
                 tab = self.current_tab
@@ -321,33 +353,58 @@ class Parameters(dict):
             if special=="header":
                 self.fprint("",tab=tab+1)
 
-    def tag_output(self, key, value):
+    def tag_output(self, key, value, collective_output=None):
 
         ### Process value ###
         if not isinstance(value,int):
             value = float(value)
 
-        ### Grab the name of the module that called this function ###
-        stack = inspect.stack()[1][0]
-        mod = inspect.getmodule(stack)
-        the_module = mod.__name__.split(".")[-1]
-        the_class = stack.f_locals["self"].__class__.__name__
-        the_method = stack.f_code.co_name
+        if self.num_procs > 1:
+            send_data = np.float64(value)
+            mpi_buff = np.zeros(self.num_procs, dtype=np.float64)
+            self.comm.Gather(send_data, mpi_buff, root=0)
 
-        ### This will tell exactly where this function was called from ###
-        # print("I was called by {}:{}.{}()".format(the_module, the_class, the_method))
+            differing_opinions = False
 
-        ### Check if that module has called before and add the dictionary entries ###
-        if the_module in self.tagged_output.keys():
-            self.tagged_output[the_module].update({key: value})
-        else:
-            self.tagged_output.update({the_module: {key: value}})
+            if not np.all(np.isclose(mpi_buff, mpi_buff[0])):
+                differing_opinions = True
 
-        ### Update the yaml file ###
-        with open(self.folder+"tagged_output.yaml","w") as file:
-            yaml.dump(self.tagged_output, file, sort_keys=False)
+            if differing_opinions or collective_output is not None:
+                if collective_output == 'sum' or 'sum' in key:
+                    value = np.sum(mpi_buff)
+                elif collective_output == 'avg' or 'avg' in key:
+                    value = np.mean(mpi_buff)
+                elif collective_output == 'max' or 'max' in key:
+                    value = np.amax(mpi_buff)
+                elif collective_output == 'min' or 'min' in key:
+                    value = np.amin(mpi_buff)
+                else:
+                    print('WARNING: tagging %s in parallel may result in disagreement between processors.' % (key))
 
-        ### Print the new dict ###
-        # print(self.tagged_output)
+                value = float(value)
+
+        if self.rank == 0:
+            ### Grab the name of the module that called this function ###
+            stack = inspect.stack()[1][0]
+            mod = inspect.getmodule(stack)
+            the_module = mod.__name__.split(".")[-1]
+            the_class = stack.f_locals["self"].__class__.__name__
+            the_method = stack.f_code.co_name
+
+            ### This will tell exactly where this function was called from ###
+            # print("I was called by {}:{}.{}()".format(the_module, the_class, the_method))
+
+            ### Check if that module has called before and add the dictionary entries ###
+            if the_module in self.tagged_output.keys():
+                self.tagged_output[the_module].update({key: value})
+            else:
+                self.tagged_output.update({the_module: {key: value}})
+
+            ### Update the yaml file ###
+            with open(self.folder+"tagged_output.yaml","w") as file:
+                yaml.dump(self.tagged_output, file, sort_keys=False)
+
+            ### Print the new dict ###
+            # print(self.tagged_output)
 
 windse_parameters = Parameters()
